[refactor] - about page 리팩토링
